+++ cedar-policy-core/src/ast/extension.rs
@@ -22,6 +22,7 @@ use crate::evaluator;
 use std::any::Any;
 use std::collections::HashMap;
 use std::fmt::{Debug, Display};
+use std::panic::{RefUnwindSafe, UnwindSafe};
 use std::sync::Arc;
 
 /// Cedar extension.
@@ -157,11 +158,11 @@ impl ExtensionFunction {
                 if args.is_empty() {
                     func()
                 } else {
-                    Err(evaluator::EvaluationError::WrongNumArguments {
-                        function_name: name.clone(),
-                        expected: 0,
-                        actual: args.len(),
-                    })
+                    Err(evaluator::EvaluationError::wrong_num_arguments(
+                        name.clone(),
+                        0,
+                        args.len(),
+                    ))
                 }
             }),
             Some(return_type),
@@ -181,11 +182,11 @@ impl ExtensionFunction {
             style,
             Box::new(move |args: &[Value]| match args.get(0) {
                 Some(arg) => func(arg.clone()),
-                None => Err(evaluator::EvaluationError::WrongNumArguments {
-                    function_name: name.clone(),
-                    expected: 1,
-                    actual: args.len(),
-                }),
+                None => Err(evaluator::EvaluationError::wrong_num_arguments(
+                    name.clone(),
+                    1,
+                    args.len(),
+                )),
             }),
             None,
             vec![arg_type],
@@ -205,11 +206,11 @@ impl ExtensionFunction {
             style,
             Box::new(move |args: &[Value]| match &args {
                 &[arg] => func(arg.clone()),
-                _ => Err(evaluator::EvaluationError::WrongNumArguments {
-                    function_name: name.clone(),
-                    expected: 1,
-                    actual: args.len(),
-                }),
+                _ => Err(evaluator::EvaluationError::wrong_num_arguments(
+                    name.clone(),
+                    1,
+                    args.len(),
+                )),
             }),
             Some(return_type),
             vec![arg_type],
@@ -231,11 +232,11 @@ impl ExtensionFunction {
             style,
             Box::new(move |args: &[Value]| match &args {
                 &[first, second] => func(first.clone(), second.clone()),
-                _ => Err(evaluator::EvaluationError::WrongNumArguments {
-                    function_name: name.clone(),
-                    expected: 2,
-                    actual: args.len(),
-                }),
+                _ => Err(evaluator::EvaluationError::wrong_num_arguments(
+                    name.clone(),
+                    2,
+                    args.len(),
+                )),
             }),
             Some(return_type),
             vec![arg_types.0, arg_types.1],
@@ -260,11 +261,11 @@ impl ExtensionFunction {
             style,
             Box::new(move |args: &[Value]| match &args {
                 &[first, second, third] => func(first.clone(), second.clone(), third.clone()),
-                _ => Err(evaluator::EvaluationError::WrongNumArguments {
-                    function_name: name.clone(),
-                    expected: 3,
-                    actual: args.len(),
-                }),
+                _ => Err(evaluator::EvaluationError::wrong_num_arguments(
+                    name.clone(),
+                    3,
+                    args.len(),
+                )),
             }),
             Some(return_type),
             vec![arg_types.0, arg_types.1, arg_types.2],
@@ -328,7 +329,7 @@ impl std::fmt::Debug for ExtensionFunction {
 /// Anything implementing this trait can be used as a first-class value in
 /// Cedar. For instance, the `ipaddr` extension uses this mechanism
 /// to implement IPAddr as a Cedar first-class value.
-pub trait ExtensionValue: Debug + Display {
+pub trait ExtensionValue: Debug + Display + Send + Sync + UnwindSafe + RefUnwindSafe {
     /// Get the name of the type of this value.
     ///
     /// Cedar has nominal typing, so two values have the same type iff they
@@ -355,7 +356,7 @@ pub struct ExtensionValueWithArgs {
 
 impl ExtensionValueWithArgs {
     /// Get the internal value
-    pub fn value(&self) -> &dyn InternalExtensionValue {
+    pub fn value(&self) -> &(dyn InternalExtensionValue) {
         self.value.as_ref()
     }
 
@@ -365,7 +366,11 @@ impl ExtensionValueWithArgs {
     }
 
     /// Constructor
-    pub fn new(value: Arc<dyn InternalExtensionValue>, args: Vec<Expr>, constructor: Name) -> Self {
+    pub fn new(
+        value: Arc<dyn InternalExtensionValue + Send + Sync>,
+        args: Vec<Expr>,
+        constructor: Name,
+    ) -> Self {
         Self {
             value,
             args,
@@ -436,7 +441,7 @@ pub trait InternalExtensionValue: ExtensionValue {
     fn cmp_extvalue(&self, other: &dyn InternalExtensionValue) -> std::cmp::Ordering;
 }
 
-impl<V: 'static + Eq + Ord + ExtensionValue> InternalExtensionValue for V {
+impl<V: 'static + Eq + Ord + ExtensionValue + Send + Sync> InternalExtensionValue for V {
     fn as_any(&self) -> &dyn Any {
         self
     }