+++ cedar-policy-core/src/ast/restricted_expr.rs
@@ -321,7 +321,8 @@ impl<'a> Hash for RestrictedExprShapeOnly<'a> {
 }
 
 /// Errors related to restricted expressions
-#[derive(Debug, Clone, PartialEq, Error)]
+
+#[derive(Debug, Clone, PartialEq, Eq, Error)]
 pub enum RestrictedExprError {
     /// An expression was expected to be a "restricted" expression, but contained
     /// a feature that is not allowed in restricted expressions. The `feature`